requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai==1.6.1
azure-ai-projects==1.0.0b7
azure-identity==1.15.0
pydantic==2.5.2
flask==2.3.3
flask-session==0.5.0